==> comparison/1. df_accession_comparison.R <==
"""
@author: ieva-balta, majocava, naaattella

"""

library(VennDiagram)

"""
This script generates a venn diagram of the accessions that are shared, and unique to each database. 

It requires the following files:
 1. PTM_old_organism.csv: Generated by running the 0.dataframe_building.py script. Contains a summary of all the accession codes in dbPTM/CPLM database from Zenodo (version 1.4)
 2. uniprot_organism_data_test.csv: Generated by running the 0.dataframe_building.py script. Contains a summary of all the accession codes in the UniProt dataset. 

"""

# Colors
light_blue = "#E4A788"
light_red = "#5361B1"

palette = c(light_blue, light_red)

# Load the data from zenodo
ptm_df <- read.csv("PTM_old_organism.csv")

# Load the uniprot csv
uniprot_df <- read.csv("uniprot_organism_data_test.csv")

# Get unique accessions with isoforms
old_accessions_iso <- unique(ptm_df$UniProt_Accession)
print(paste("Number of final proteins with isoforms (dbptm/cplm):", length(old_accessions_iso)))

uniprot_accessions_iso <- unique(uniprot_df$UniProt_Accession)

# Compute union
total_unique_iso <- union(old_accessions_iso, uniprot_accessions_iso)

# Compute intersection
common_iso <- intersect(old_accessions_iso, uniprot_accessions_iso)

# PTMs unique in dbptm + CPLM
unique_old_iso <- setdiff(old_accessions_iso, uniprot_accessions_iso)

# PTMs unique in uniprot
unique_uniprot_iso <- setdiff(uniprot_accessions_iso, old_accessions_iso)

x <- list(`UniProt` = uniprot_accessions_iso, `dbPTM/CPLM` = old_accessions_iso)

venn_plot = venn.diagram(
  x,
  filename = NULL,
  fill = palette,
  alpha = 0.5,
  cex = 0.9,
  fontfamily = "Arial",
  fontface = "bold",
  cat.cex = 0,
  margin = 0.2,              # prevents clipping
  #  cat.pos = c(0, 0),      # dbPTM label (left), UniProt label (right)
  #  cat.dist = c(0.05, 0.05),
  #  cat.fontfamily = "Arial",
  lty= "blank",
  lwd = 0, 
  rotation.degree = 180,
  bg = NULL
)
svg("Final_pictures/accession_venn_diagram_set.svg")
grid.draw(venn_plot)
dev.off()

# Create a Venn diagram
venn.diagram(
  x = list(UniProt = uniprot_accessions_iso, `CPLM/dbPTM` = old_accessions_iso),
  filename = "accessions_venn_iso.png",
  main = "Venn Diagram of PTMs with Isoforms"
)
