==> comparison/4. df_ptm_comparison.R <==
"""
@author: ieva-balta, majocava, naaattella
"""

library(dplyr)
library(ggplot2)
library(readr)
library(ggthemes)
library(VennDiagram)
library(stringr)
library(tidyr)
library(Biostrings)

"""
This script compares the accession in both dbPTM/CPLM and UniProt versions in FLAMS
Old refers to the dbPTM/CPLM version

It requires the file: 
 1. PTM_old_organism.csv: Generated by running the 0.dataframe_building.py script. Contains a summary of all the accession codes in dbPTM/CPLM database from Zenodo (version 1.4)
 2. For the UniProt data, all the csv files are required to get the PTM types. 

"""

# Colorblind-friendly palette
cb_palette = colorblind_pal()(2)

light_blue = "#E4A788"
light_red = "#5361B1"

palette = c(light_blue, light_red)

# Change the working directory to your working directoy
setwd("C:/Users/majoc/Downloads/OneDrive_1_11-8-2025")

# Create dataframe for the UniProt database

uniprot_folder = "Ebi_uniprot/dbs_clean/dbs_clean/"
skip_files = c("_deduplicated_records-2.0.csv", "_sequences-2.0.csv", "classified-2.0.csv")

files = list.files(uniprot_folder, pattern = "\\.csv$", full.names = TRUE)
files = files[!basename(files) %in% skip_files]

all_dfs = lapply(files, function(file) {
  df = read_csv(file, show_col_types = FALSE)
  ptm_type = basename(file) %>% str_replace("-2.0.csv", "") %>% str_replace_all("_", " ")
  df %>% mutate(PTM_type = ptm_type, PTM_count = 1)
})

combined_df = bind_rows(all_dfs)

combined_df = combined_df %>% distinct(Accession, Position, PTM_type, .keep_all = TRUE)
write_csv(combined_df, "combined_ptms_uniprot_deduplicate.csv")

# Load UniProt combined
uniprot_df = read_csv("combined_ptms_uniprot_deduplicate.csv", show_col_types = FALSE)

# Load dbPTM/CPLM data, change the path if needed. 
dbptm = read_csv("PTM_old_organism.csv", show_col_types = FALSE)

# Deduplicate the old
old_ptm = dbptm %>% distinct(UniProt_Accession, Position, Modification, .keep_all = TRUE)

# Save deduplicated dbPTM/CPLM
write_csv(old_ptm, "Final_pictures/deduplicated_ptms_zenodo.csv")

old_ptm = read_csv("Final_pictures/deduplicated_ptms_zenodo.csv", show_col_types = FALSE)

# UniProt accessions
old_accessions = unique(old_ptm$UniProt_Accession)
uniprot_accessions = unique(uniprot_df$Accession)

# Compute sets
total_unique = union(old_accessions, uniprot_accessions)
common = intersect(old_accessions, uniprot_accessions)
unique_old = setdiff(old_accessions, uniprot_accessions)
unique_uniprot = setdiff(uniprot_accessions, old_accessions)

# Create category dataframe
all_accessions = total_unique
categories = ifelse(all_accessions %in% common, "both",
                     ifelse(all_accessions %in% unique_old, "dbPTM/CPLM", "UniProt"))
category_df = data.frame(Accession = all_accessions, Category = categories)
#write_csv(category_df, "Final_pictures/accession_categories.csv")

# Add Accession column for filtering PTM counts
old_ptm = old_ptm %>% mutate(Accession = UniProt_Accession)

# PTM counts for dbPTM/CPLM and UniProt
ptm_counts_dbptm_cplm = old_ptm %>%
  mutate(category = ifelse(Accession %in% common, "both", "dbPTM/CPLM")) %>%
  group_by(category) %>%
  summarise(count = n())

ptm_counts_uniprot = uniprot_df %>%
  mutate(category = ifelse(Accession %in% common, "both", "UniProt")) %>%
  group_by(category) %>%
  summarise(count = n())

ptms_unique_dbptm_cplm = ptm_counts_dbptm_cplm %>% filter(category == "dbPTM/CPLM") %>% pull(count) %>% .[1] %||% 0
ptms_shared_dbptm_cplm = ptm_counts_dbptm_cplm %>% filter(category == "both") %>% pull(count) %>% .[1] %||% 0
ptms_unique_uniprot = ptm_counts_uniprot %>% filter(category == "UniProt") %>% pull(count) %>% .[1] %||% 0
ptms_shared_uniprot = ptm_counts_uniprot %>% filter(category == "both") %>% pull(count) %>% .[1] %||% 0

print(paste("ptms_unique_dbptm_cplm:", ptms_unique_dbptm_cplm))
print(paste("ptms_shared_dbptm_cplm:", ptms_shared_dbptm_cplm))
print(paste("ptms_unique_uniprot:", ptms_unique_uniprot))
print(paste("ptms_shared_uniprot:", ptms_shared_uniprot))

# Graph of venn diagram based on the accession/position/type from the superdeduplicated files
uniprot_df = uniprot_df %>% mutate(PTM_type = tolower(PTM_type))
old_ptm = old_ptm %>% mutate(Modification = tolower(Modification))

uniprot_ptms = with(uniprot_df, paste(Accession, Position, PTM_type, sep = "|"))
dbptm_ptms = with(old_ptm, paste(Accession, Position, Modification, sep = "|"))

only_uniprot = setdiff(uniprot_ptms, dbptm_ptms)
only_dbptm = setdiff(dbptm_ptms, uniprot_ptms)
shared_ptms = intersect(uniprot_ptms, dbptm_ptms)

x = list(`UniProt` = uniprot_ptms, `dbPTM/CPLM` = dbptm_ptms)

venn_plot = venn.diagram(
  x,
  filename = NULL,
  fill = palette,
  alpha = 0.5,
  cex = 0.9,
  fontfamily = "Arial",
  fontface = "bold",
  cat.cex = 0,
  margin = 0.2,              # prevents clipping
#  cat.pos = c(0, 0),      # dbPTM label (left), UniProt label (right)
#  cat.dist = c(0.05, 0.05),
#  cat.fontfamily = "Arial",
  lty= "blank",
  lwd = 0, 
  rotation.degree = 180,
  bg = NULL
)
svg("Final_pictures/ptm_venn_diagram_set.svg")
grid.draw(venn_plot)
dev.off()
ggsave("Final_pictures/ptm_venn_diagram_set.svg", plot = venn_plot, device = "svg")

# Compute PTM type counts in both databases for every accession
old_ptm_full = old_ptm
uniprot_df_full = uniprot_df

ptm_type_counts_dbptm_cplm_full = old_ptm_full %>%
  group_by(Modification) %>%
  summarise(count = n()) %>%
  arrange(desc(count))

ptm_type_counts_uniprot_full = uniprot_df_full %>%
  group_by(PTM_type) %>%
  summarise(count = n()) %>%
  arrange(desc(count))

all_ptm_types_full = sort(union(ptm_type_counts_dbptm_cplm_full$Modification, ptm_type_counts_uniprot_full$PTM_type))

aligned_counts_full = data.frame(
  ptm_type = all_ptm_types_full,
  `dbPTM/CPLM` = sapply(all_ptm_types_full, function(ptm) {
    ptm_type_counts_dbptm_cplm_full %>% filter(Modification == ptm) %>% pull(count) %>% .[1] %||% 0
  }),
  UniProt = sapply(all_ptm_types_full, function(ptm) {
    ptm_type_counts_uniprot_full %>% filter(PTM_type == ptm) %>% pull(count) %>% .[1] %||% 0
  })
)

aligned_counts_full = aligned_counts_full %>%
  arrange(desc(UniProt))

aligned_long_full = aligned_counts_full %>%
  pivot_longer(cols = c(`dbPTM.CPLM`, UniProt), values_to = "count") %>%
  mutate(ptm_type = factor(ptm_type, levels = aligned_counts_full$ptm_type))

write_csv(aligned_counts_full, "Final_pictures/aligned_ptm_counts_full.csv")

###################################################################
############## Compare broad vs specific PTM types ################
###################################################################

create_vector = function(path) {
  fasta = readAAStringSet(path)
  headers = names(fasta)
  entries = c()
  for (header in headers) {
    parts = strsplit(header, " ")[[1]]
    seq_id = parts[1]
    id_parts = strsplit(seq_id, "\\|")[[1]]
    up_id = id_parts[1]
    pos = id_parts[2]
    if (grepl("_", up_id, fixed = TRUE)) {
      up_id = strsplit(up_id, "_", fixed = TRUE)[[1]][1]
    }
    entry = paste(up_id, pos, sep = "|")
    entries = c(entries, entry)
  }
  return(entries)
}

compare = function(vec1, vec2) {
  common = intersect(vec1, vec2)
  return(list(common, length(common)))
}

# Change the PTM for other comparisons

dis_bond = create_vector("disulfide_bond-2.0.fasta")
s_nitro = create_vector("s-nitrosylation-1.4.fasta")
print(paste("s-nitro in disulfide bond:", compare(s_nitro, dis_bond)))


#########################################################
################# Graphs: Dot Plot ######################
#########################################################

# Plot a venn diagram per PTM type for the top 5 PTM types in both databases
top_ptm_types_full = aligned_counts_full$ptm_type[1:5]
for (ptm_type in top_ptm_types_full) {
  uniprot_subset = uniprot_df_full %>%
    filter(PTM_type == ptm_type) %>%
    select(Accession, Position) %>%
    distinct() %>%
    with(paste(Accession, Position, sep = "|"))
  
  dbptm_subset = old_ptm_full %>%
    filter(Modification == ptm_type) %>%
    select(Accession, Position) %>%
    distinct() %>%
    with(paste(Accession, Position, sep = "|"))
  
  x = list(`UniProt PTMs` = uniprot_subset, `dbPTM/CPLM PTMs` = dbptm_subset)
  venn.diagram(x, filename = paste0("Final_pictures/venn_diagram_full_", gsub(" ", "_", ptm_type), ".png"),
               fill = cb_palette, alpha = 0.5, cex = 1.5, cat.cex = 1.5,
               main = paste("Venn Diagram for PTM Type:", ptm_type))
}

# Dot Plot

df = read.csv("aligned_ptm_counts_full.csv")

# make the column names explicit
df = df %>%
  rename(
    PTM_type   = X,           # PTM name
    dbPTM_CPLM = dbPTM.CPLM,  # dbPTM column
    UniProt    = UniProt      # UniProt column
  )
df_long = df %>%
  mutate(
    dbPTM_log = log10(dbPTM_CPLM + 1),
    UniProt_log = log10(UniProt + 1)
  ) %>%
  select(PTM_type, dbPTM_log, UniProt_log) %>%
  pivot_longer(
    cols = ends_with("_log"),
    names_to = "database",
    values_to = "log_count"
  ) %>%
  mutate(
    database = recode(database,
                      "dbPTM_log" = "dbPTM.CPLM",
                      "UniProt_log" = "UniProt")
  )


# order by uniprot
ptm_order = df %>%
  arrange(desc(UniProt)) %>%      # highest -> lowest
  pull(PTM_type)
df_long = df_long %>%
  mutate(
    PTM_type = factor(PTM_type, levels = ptm_order)
  )

p = ggplot(df_long, aes(x = PTM_type, y = log_count, color = database)) +
  geom_line(aes(group = PTM_type), color="grey70", size = 1.2) +
  geom_point(size = 5, alpha = 0.85) +
  scale_color_manual(values = palette,
                     labels = c(
                       "dbPTM.CPLM" = "dbPTM/CPLM",
                       "UniProt"    = "UniProt"
                     )) + 
  labs(
    y = "Number of PTMs (log10-scale)",
    color = NULL,
  ) +
  theme_minimal(base_size = 10) +
  theme(panel.grid = element_blank(),
        text = element_text(family = "Arial"),
        axis.text.x = element_text(angle=90, hjust=1, size = 14),
        axis.text.y = element_text(size = 14),
        legend.position = c(0.95, 0.95),     # << move here,
        legend.key = element_rect(fill = "white", color = NA),
        legend.text = element_text(size = 12),
        legend.key.size = unit(0.8, "cm"),
        panel.background  = element_rect(fill = NA, colour = NA),   # key
        plot.background   = element_rect(fill = NA, colour = NA),   # key
        legend.background = element_rect(fill = NA, colour = NA),
        # keep axes
        axis.line         = element_line(color = "black", linewidth = 0.5),
        axis.ticks        = element_line(color = "black"),
        axis.title.y = element_text(size = 16),
        axis.title.x = element_blank()
  )
p


ggsave(
  "ptm_dotplot.svg",
  p,
  width = 24,
  height = 7,
  bg = "transparent"

)
